+++ arq/testing.py
@@ -8,6 +8,7 @@ See arq's own tests for examples of usage.
 """
 import asyncio
 import logging
+from contextlib import contextmanager
 from datetime import datetime, timedelta
 
 from .utils import RedisMixin, timestamp
@@ -28,6 +29,11 @@ class RaiseWorker(BaseWorker):
         raise RuntimeError(msg)
 
 
+@contextmanager
+def redis_context_manager(r):
+    yield r
+
+
 class MockRedis:
     """
     Very simple mock of aioredis > Redis which allows jobs to be enqueued and executed without
@@ -37,7 +43,12 @@ class MockRedis:
         self.loop = loop or asyncio.get_event_loop()
         self.data = {} if data is None else data
         self._expiry = {}
+        self._pool_or_conn = type('MockConnection', (), {'_loop': self.loop})
         logger.info('initialising MockRedis, data id: %s', None if data is None else id(data))
+
+    def __await__(self):
+        return redis_context_manager(self)
+        yield  # pragma: no cover
 
     async def rpush(self, list_name, data):
         logger.info('rpushing %s to %s', data, list_name)
@@ -92,59 +103,30 @@ class MockRedis:
     async def get(self, key):
         return self._get(key)
 
-
-class MockRedisPoolContextManager:
-    def __init__(self, loop, data):
-        self.loop = loop
-        self.data = data
-
-    async def __aenter__(self):
-        return MockRedis(loop=self.loop, data=self.data)
-
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
-        pass
-
-
-class MockRedisPool:
-    def __init__(self, loop=None):
-        self._loop = loop or asyncio.get_event_loop()
-        self.data = {}
-
-    async def acquire(self):
-        return MockRedis(loop=self._loop, data=self.data)
-
-    def release(self, conn):
-        pass
-
-    def get(self):
-        return MockRedisPoolContextManager(self._loop, self.data)
-
     def close(self):
         pass
 
     async def wait_closed(self):
         pass
 
-    async def clear(self):
-        self.data = {}
-
 
 class MockRedisMixin(RedisMixin):
     """
     Dependent of RedisMixin which uses MockRedis rather than real redis to enqueue jobs.
     """
+
     async def create_redis_pool(self):
-        return self.redis_pool or MockRedisPool(self.loop)
+        return self.redis or MockRedis(loop=self.loop)
 
     @property
     def mock_data(self):
-        self.redis_pool = self.redis_pool or MockRedisPool(self.loop)
-        return self.redis_pool.data
+        self.redis = self.redis or MockRedis(loop=self.loop)
+        return self.redis.data
 
     @mock_data.setter
     def mock_data(self, data):
-        self.redis_pool = self.redis_pool or MockRedisPool(self.loop)
-        self.redis_pool.data = data
+        self.redis = self.redis or MockRedis(loop=self.loop)
+        self.redis.data = data
 
 
 class MockRedisWorker(MockRedisMixin, BaseWorker):