+++ arq/worker.py
@@ -136,7 +136,7 @@ class BaseWorker(RedisMixin):
             redis_settings=self.redis_settings,
             worker=self,
             loop=self.loop,
-            existing_pool=await self.get_redis_pool(),
+            existing_redis=await self.get_redis(),
         )
 
     @classmethod
@@ -202,7 +202,7 @@ class BaseWorker(RedisMixin):
         work_logger.info('shadows: %s | queues: %s', self.shadow_names, ', '.join(self.queues))
 
         self.drain = self.drain_class(
-            redis_pool=await self.get_redis_pool(),
+            redis=await self.get_redis(),
             max_concurrent_tasks=self.max_concurrent_tasks,
             shutdown_delay=self.shutdown_delay - 1,
             timeout_seconds=self.timeout_seconds,
@@ -266,13 +266,13 @@ class BaseWorker(RedisMixin):
 
     async def _check_health(self):
         r = 1
-        async with await self.get_redis_conn() as redis:
-            data = await redis.get(self.health_check_key)
-            if not data:
-                work_logger.warning('Health check failed: no health check sentinel value found')
-            else:
-                work_logger.info('Health check successful: %s', data.decode())
-                r = 0
+        redis = await self.get_redis()
+        data = await redis.get(self.health_check_key)
+        if not data:
+            work_logger.warning('Health check failed: no health check sentinel value found')
+        else:
+            work_logger.info('Health check successful: %s', data.decode())
+            r = 0
         # only need to close redis not deal with queues etc., hence super close
         await super().close()
         return r