+++ tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from aioredis import create_pool, create_redis
+from aioredis import create_redis, create_redis_pool
 
 from .fixtures import DemoActor, MockRedisDemoActor, MockRedisWorker
 
@@ -27,26 +27,24 @@ def redis_conn(loop):
 
 
 @pytest.yield_fixture
-def redis_pool(loop):
+def redis(loop):
     """
     yield fixture which creates a redis connection, and flushes redis before the test.
 
     Note: redis is not flushed after the test both for performance and to allow later debugging.
     """
-    async def _create_pool():
-        p = await create_pool(('localhost', 6379), loop=loop)
-        async with p.get() as redis:
-            await redis.flushall()
-        return p
+    async def _create_redis():
+        r = await create_redis_pool(('localhost', 6379), loop=loop)
+        await r.flushall()
+        return r
 
-    async def _close_pool(p):
-        p.close()
-        await p.wait_closed()
-        await p.clear()
+    async def _close(r):
+        r.close()
+        await r.wait_closed()
 
-    pool = loop.run_until_complete(_create_pool())
-    yield pool
-    loop.run_until_complete(_close_pool(pool))
+    redis_ = loop.run_until_complete(_create_redis())
+    yield redis_
+    loop.run_until_complete(_close(redis_))
 
 
 @pytest.yield_fixture